fixes and optimizations

+++ src/graphics/ui/elements/control/TextBox.cpp
@@ -333,7 +333,7 @@ void TextBox::click(GUI*, int x, int y) {
 }
 
 void TextBox::mouseMove(GUI*, int x, int y) {
-    size_t index = calcIndexAt(x, y);
+    ptrdiff_t index = calcIndexAt(x, y);
     setCaret(index);
     extendSelection(index);
     resetMaxLocalCaret();
@@ -416,7 +416,7 @@ void TextBox::stepDefaultUp(bool shiftPressed, bool breakSelection) {
         uint offset = std::min(size_t(maxLocalCaret), getLineLength(caretLine - 1) - 1);
         setCaret(label->getTextLineOffset(caretLine - 1) + offset);
     } else {
-        setCaret(0ULL);
+        setCaret(static_cast<size_t>(0));
     }
     if (shiftPressed) {
         if (selectionStart == selectionEnd) {
@@ -564,7 +564,7 @@ size_t TextBox::getCaret() const {
     return caret;
 }
 
-void TextBox::setCaret(ssize_t position) {
+void TextBox::setCaret(ptrdiff_t position) {
     if (position < 0) {
         setCaret(static_cast<size_t>(input.length() + position + 1));
     } else {
@@ -593,7 +593,7 @@ void TextBox::setCaret(size_t position) {
     if (realoffset-width > 0) {
         setTextOffset(textOffset + realoffset - width);
     } else if (realoffset < 0) {
-        setTextOffset(std::max(textOffset + realoffset, 0ULL));
+        setTextOffset(std::max(textOffset + realoffset, static_cast<size_t>(0)));
     }
 }
